% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{center_scale}
\alias{center_scale}
\title{Center a numeric continuous scale}
\usage{
center_scale(
  data,
  scale_var,
  name = paste0(as_label(expr({
     {
         scale_var
     }
 })), "_centered")
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} or a \link{tdcmm} model}

\item{scale_var}{a numeric variable for which the scale gets centered}

\item{name}{the name of the new variable. By default, this is the same name
as \code{scale_var} but suffixed with \verb{_centered}.}
}
\value{
a \link{tdcmm} model
}
\description{
Subtracts the mean from each individual data point. A centered scale then
centers at a mean of 0.0.
}
\examples{
WoJ \%>\% center_scale(autonomy_emphasis)
WoJ \%>\% center_scale(autonomy_emphasis, name = "my_centered_variable")

WoJ \%>\%
  center_scale(autonomy_emphasis) \%>\%
  tab_frequencies(autonomy_emphasis, autonomy_emphasis_centered)
}
\seealso{
Other scaling: 
\code{\link{change_scale}()},
\code{\link{reverse_scale}()},
\code{\link{z_scale}()}
}
\concept{scaling}
