% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{change_scale}
\alias{change_scale}
\title{Change/Rescale a numeric continuous scale}
\usage{
change_scale(
  data,
  scale_var,
  change_to_min = 0,
  change_to_max = 1,
  name = paste0(as_label(expr({
     {
         scale_var
     }
 })), "_",
    change_to_min, "to", change_to_max)
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} or a \link{tdcmm} model}

\item{scale_var}{a numeric variable for which the scale gets changed}

\item{change_to_min}{numeric minimum of the new range (default is \code{0})}

\item{change_to_max}{numeric maximum of the new range (default is \code{1})}

\item{name}{the name of the new variable. By default, this is the same name
as \code{scale_var} but suffixed with \code{change_to_min} and \code{change_to_max} so that,
for example, "variable" becomes "variable_3to5".}
}
\value{
a \link{tdcmm} model
}
\description{
Given a specified minimum and maximum, this function translates each value
into a new value within this specified range, thereby keeping distances the
same. Hence, mean and standard deviations change but. However, if both the
original scale and the changed scale were to be z-standardized, they would
turn out equal again.
}
\examples{
WoJ \%>\% change_scale(autonomy_emphasis)
WoJ \%>\% change_scale(autonomy_emphasis, change_to_min = 1, change_to_max = 10, name = "my_changed_variable")

WoJ \%>\%
  change_scale(autonomy_emphasis) \%>\%
  tab_frequencies(autonomy_emphasis, autonomy_emphasis_0to1)
}
\seealso{
Other scaling: 
\code{\link{center_scale}()},
\code{\link{reverse_scale}()},
\code{\link{z_scale}()}
}
\concept{scaling}
