% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_agreement.R
\name{check_equal}
\alias{check_equal}
\title{Check if all values in a vector are the same}
\usage{
check_equal(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Checks if all values in a vector are the same
}
