% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{cohens_d}
\alias{cohens_d}
\title{Compute Cohen's d}
\usage{
cohens_d(x, y, pooled_sd = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{pooled_sd}{a logical indicating whether to use the pooled standard
deviation in the calculation of Cohen's d. Defaults to \code{TRUE}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
a \code{dbl}
}
\description{
Computes the effect size estimate Cohen's d for two sets of numerical values
}
\seealso{
Other t-test: \code{\link{compute_t_test}},
  \code{\link{t_test}}
}
\concept{t-test}
