% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{col_percs}
\alias{col_percs}
\title{Compute column percentages}
\usage{
col_percs(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
a \code{dbl}
}
\description{
Computes column percentages
}
