% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unianova.R
\name{compute_aov}
\alias{compute_aov}
\title{Compute one-way ANOVA}
\usage{
compute_aov(test_var, data, group_var, descriptives, post_hoc)
}
\arguments{
\item{test_var}{Test variable}

\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{group_var}{group variable (column name)}

\item{descriptives}{a logical indicating whether descriptive statistics (mean
& standard deviation) for all group levels should be added to the returned
tibble. Defaults to \code{FALSE}.}

\item{post_hoc}{a logical indicating whether post-hoc tests (Tukey's HSD)
should be computed. Results of the post-hoc test will be added in a list
column of result tibbles.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Computes and outputs a one-way ANOVA for one test variable
}
\seealso{
Other ANOVA: \code{\link{unianova}}
}
\concept{ANOVA}
