% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr.R
\name{compute_icr}
\alias{compute_icr}
\title{Compute intercoder reliability estimates}
\usage{
compute_icr(data, unit_var, coder_var, ..., levels = NULL,
  na.omit = FALSE, agreement = TRUE, holsti = TRUE,
  kripp_alpha = TRUE)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{unit_var}{Variable with unit identifiers}

\item{coder_var}{Variable with coder identifiers}

\item{...}{Variables to compute intercoder reliability estimates for}

\item{na.omit}{Logical indicating whether \code{NA} values should be stripped
before computation. Defaults to FALSE.}

\item{agreement}{Logical indicating whether simple percent agreement should
be computed. Defaults to TRUE.}

\item{holsti}{Logical indicating whether Holsti's reliability estimate
(mean pairwise agreement) should be computed. Defaults to TRUE.}

\item{kripp_alpha}{Logical indicating whether Krippendorff's Alpha should
be computed. Defaults to TRUE.}

\item{...}{Optional named vector with levels of test variables}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Computes various intercoder reliability estimates
}
\seealso{
Other intercoder reliability: \code{\link{icr_agreement}},
  \code{\link{icr_holsti}}, \code{\link{icr_kripp_alpha}},
  \code{\link{icr_test}}, \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
