% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{compute_reliability}
\alias{compute_reliability}
\title{Compute reliability}
\usage{
compute_reliability(data, ..., type = "alpha", interval.type = NULL,
  bootstrap.samples = NULL, conf.level = NULL)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{...}{Variables to compute reliability estimate for}

\item{type}{See \code\link[MBESS::ci.reliability]{MBESS::ci.reliability}}

\item{interval.type}{See \code\link[MBESS::ci.reliability]{MBESS::ci.reliability}}

\item{bootstrap.samples}{See \code\link[MBESS::ci.reliability]{MBESS::ci.reliability}}

\item{conf.level}{See \code\link[MBESS::ci.reliability]{MBESS::ci.reliability}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Computes reliability estimates of index/scale variables.
Provides a wrapper for \code\link[MBESS::ci.reliability]{MBESS::ci.reliability}.
}
\concept{reliability}
