% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_partial.R
\name{correlate_partial}
\alias{correlate_partial}
\title{Compute partial correlation coefficients for three variables}
\usage{
correlate_partial(data, ..., method = "pearson")
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{...}{three variables to compute partial correlations for (column names).}

\item{method}{a character string indicating which partial correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or "spearman"}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Computes pairwise partial correlations (pearson, kendall, spearman) for
all combinations, i.e. pairs, of specified variables while controlling for the third variable.
}
\examples{
WoJ \%>\% correlate_partial(ethics_1, ethics_2, ethics_3)

}
\seealso{
Other correlations: 
\code{\link{correlate}()}
}
\concept{correlations}
