% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation_test}
\alias{correlation_test}
\title{Compute correlation test}
\usage{
correlation_test(var_comb, data, method)
}
\arguments{
\item{var_comb}{A character vector containing the name of two variables}

\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{method}{a character string indicating which correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or "spearman"}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Computes a correlation test for a two variables
}
\seealso{
Other correlations: \code{\link{correlate}}
}
\concept{correlations}
