% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kripp_alpha.R
\name{count_value_in_unit}
\alias{count_value_in_unit}
\title{Count specific value in unit}
\usage{
count_value_in_unit(val, tab_u)
}
\arguments{
\item{val}{Value to count}

\item{tab_u}{A frequency table of values in the unit}
}
\description{
Counts occurences of a specific value in a single unit of a units-coders matrix.
}
