% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{cramer_V}
\alias{cramer_V}
\title{Compute Cramer's V}
\usage{
cramer_V(chi2)
}
\arguments{
\item{chi2}{Output from a \code{chisq.test()}.}
}
\value{
a \code{dbl}
}
\description{
Computes Cramer's V
}
\seealso{
Other categorical: \code{\link{crosstab}},
  \code{\link{tab_frequencies}}
}
\concept{categorical}
