% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kripp_alpha.R
\name{delta_sq}
\alias{delta_sq}
\title{Compute value of difference function}
\usage{
delta_sq(c, k, vum, var_level)
}
\arguments{
\item{c}{Value c}

\item{k}{Value k}

\item{vum}{A values-units matrix}

\item{var_level}{Variable level}
}
\description{
Computes value of the difference function
}
\references{
Krippendorff, K. (2011). Computing Krippendorff's Alpha-Reliability.
Retrieved from http://repository.upenn.edu/asc_papers/43
}
