% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_groups}
\alias{describe_groups}
\title{Describe groups by variable}
\usage{
describe_groups(data, var, ..., na.rm = TRUE)
}
\arguments{
\item{data}{A (tidy) dataset}

\item{var}{Variable to describe by (column name)}

\item{...}{Variable(s) to group by (column name)}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds. Defaults to TRUE.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Describe one or more groups by descriptive statistics for one continous
variable.
}
\examples{
describe(mtcars, mpg, cyl, am)

}
