% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-designs.R
\name{design_grey}
\alias{design_grey}
\title{Grey design}
\usage{
design_grey()
}
\value{
a list with \code{main_color_1}, a vector of 12 \code{main_colors}, a
corresponding \code{main_contrast_1} (the color of text to write on top of the
main color) and a corresponding \code{main_contrasts}, the \code{main_size} (for
lines), a \code{comparison_linetype}, \code{comparison_color}, and \code{comparison_size}
for all lines that act as comparative lines, and a \link[ggplot2:theme]{theme()}
}
\description{
Grey design
}
