% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab_vars}
\alias{grab_vars}
\title{Keep existing variables or get all numeric variables}
\usage{
grab_vars(data, vars)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{vars}{Variables passed to function with \code{...}, wrapped in \code{rlang::enquos}}
}
\value{
Variables as symbols
}
\description{
Keeps existing variables if they were specified in the function call or
gets all numeric variables.
}
