% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_agreement.R
\name{icr_agreement}
\alias{icr_agreement}
\title{Compute simple percent agreement}
\usage{
icr_agreement(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes simple percent agreement for a units-coders matrix
}
