% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kappa.R
\name{icr_cohens_kappa}
\alias{icr_cohens_kappa}
\title{Compute Cohen's Kappa}
\usage{
icr_cohens_kappa(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes Cohen's Kappa.
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}}, \code{\link{icr_holsti}},
  \code{\link{icr_kripp_alpha}}, \code{\link{icr_test}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
