% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kappa.R
\name{icr_fleiss_kappa}
\alias{icr_fleiss_kappa}
\title{Compute Fleiss' Kappa}
\usage{
icr_fleiss_kappa(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes Fleiss' Kappa.
}
\references{
Fleiss, J. L. (1971). Measuring nominal scale agreement among many raters.
Psychological Bulletin, 76(5), 378-382. https://doi.org/10.1037/h0031619
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}},
  \code{\link{icr_brennan_prediger}},
  \code{\link{icr_cohens_kappa}},
  \code{\link{icr_holstis_CR}},
  \code{\link{icr_kripp_alpha}}, \code{\link{test_icr}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
