% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_holsti.R
\name{icr_holsti}
\alias{icr_holsti}
\title{Compute Holsti's reliability estimate}
\usage{
icr_holsti(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes Holsti's reliability estimate (mean pairwise agreement).
}
