% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_agreement.R
\name{icr_holstis_CR}
\alias{icr_holstis_CR}
\title{Compute Holsti's reliability estimate}
\usage{
icr_holstis_CR(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes Holsti's reliability estimate (mean pairwise agreement).
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}},
  \code{\link{icr_brennan_prediger}},
  \code{\link{icr_cohens_kappa}},
  \code{\link{icr_fleiss_kappa}},
  \code{\link{icr_kripp_alpha}}, \code{\link{test_icr}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
