% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kripp_alpha.R
\name{icr_kripp_alpha}
\alias{icr_kripp_alpha}
\title{Compute Krippendorff's Alpha}
\usage{
icr_kripp_alpha(ucm, var_level = NULL)
}
\arguments{
\item{ucm}{A units-coders matrix}

\item{var_level}{Variable level}
}
\description{
Computers Krippendorff's Alpha intercoder reliability estimate as detailed
in Krippendorff (2011), section E.
}
\references{
Krippendorff, K. (2011). Computing Krippendorff's Alpha-Reliability.
Retrieved from http://repository.upenn.edu/asc_papers/43
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}},
  \code{\link{icr_brennan_prediger}},
  \code{\link{icr_cohens_kappa}},
  \code{\link{icr_fleiss_kappa}},
  \code{\link{icr_holstis_CR}}, \code{\link{test_icr}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
