% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr.R
\name{icr_test}
\alias{icr_test}
\title{Compute intercoder reliability estimates for one test variable}
\usage{
icr_test(test_var, data, unit_var, coder_var, levels = c(), na.omit,
  agreement, holsti, kripp_alpha)
}
\arguments{
\item{test_var}{Variable to compute estimates for}

\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{unit_var}{Variable with unit identifiers}

\item{coder_var}{Variable with coder identifiers}

\item{levels}{Optional named vector with levels of test variables}

\item{na.omit}{Logical indicating whether \code{NA} values should be stripped
before computation. Defaults to FALSE.}

\item{agreement}{Logical indicating whether simple percent agreement should
be computed. Defaults to TRUE.}

\item{holsti}{Logical indicating whether Holsti's reliability estimate
(mean pairwise agreement) should be computed. Defaults to TRUE.}

\item{kripp_alpha}{Logical indicating whether Krippendorff's Alpha should
be computed. Defaults to TRUE.}
}
\description{
Computes intercoder reliability estimates for one test variable
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}},
  \code{\link{icr_cohens_kappa}}, \code{\link{icr_holsti}},
  \code{\link{icr_kripp_alpha}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
