% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kripp_alpha.R
\name{kalpha_denom}
\alias{kalpha_denom}
\title{Compute Krippendorffs's Alpha denominator}
\usage{
kalpha_denom(vum, numvals, var_level)
}
\arguments{
\item{vum}{Values-units matrix}

\item{numvals}{Amount of unique values in codings (categories)}

\item{var_level}{Variable level}
}
\description{
Computes the denominator in Krippendorff's Alpha formula
}
\references{
Krippendorff, K. (2011). Computing Krippendorff's Alpha-Reliability.
Retrieved from http://repository.upenn.edu/asc_papers/43
}
