% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kripp_alpha.R
\name{kalpha_num_values}
\alias{kalpha_num_values}
\title{Compute Krippendorffs's Alpha numerator values}
\usage{
kalpha_num_values(u, numvals, vum, var_level)
}
\arguments{
\item{u}{Unit of a values-units matrix}

\item{numvals}{Amount of unique values in codings (categories)}

\item{vum}{Values-units matrix}

\item{var_level}{Variable level}
}
\description{
Computes the values in the numerator in Krippendorff's Alpha formula
}
\references{
Krippendorff, K. (2011). Computing Krippendorff's Alpha-Reliability.
Retrieved from http://repository.upenn.edu/asc_papers/43
}
