% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_crosstabs.R
\name{knit_crosstab}
\alias{knit_crosstab}
\title{knit crosstab table variables}
\usage{
knit_crosstab(
  data,
  row_var,
  col_var,
  width_pct = 80,
  weight = NULL,
  na = FALSE,
  percent = TRUE,
  num_decimal = 0,
  percent_decimal = 0,
  chi_square = TRUE,
  name_row_total = "Gesamt",
  name_percent = "Prozent",
  cums = TRUE
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{row_var}{the row variable.}

\item{col_var}{the column variable.}

\item{weight}{Weighting variable. Default = NULL, means all weights are 1.}

\item{na}{= FALSE include NA. Default is FALSE.}

\item{num_decimal}{Decimal places of the numeric columns. Default is 0.}

\item{percent_decimal}{Decimal places of the percent columns. Default is 0.}

\item{chi_square}{= FALSE include NA. Default is FALSE.}

\item{name_row_total}{the name of the "Total" column.}

\item{name_percent}{= "Prozent" Decimal places of the percent columns. Default is 0.}

\item{cums}{= FALSE Decimal places of the percent columns. Default is 0.}
}
\value{
a gt-table
}
\description{
Knits table for one ore more variables
}
\examples{
WoJ |> knit_crosstab(row_var = employment, col_var = reach)
WoJ |> knit_crosstab(row_var = employment, col_var = reach, num_decimal = 1, percent_decimal = 1, cums = FALSE, name_percent = "Percent")

}
\seealso{
Other categorical: 
\code{\link{crosstab}()},
\code{\link{knit_frequencies}()},
\code{\link{tab_frequencies}()}
}
\concept{categorical}
