% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_frequencies.R
\name{knit_frequencies}
\alias{knit_frequencies}
\title{knit frequencies table variables}
\usage{
knit_frequencies(
  data,
  ...,
  weight = NULL,
  num_decimal = 0,
  percent_decimal = 0,
  cums = TRUE
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{...}{List of variables.}

\item{weight}{Weighting variable. Default = NULL, means all weights are 1.}

\item{num_decimals}{Decimal places of the numeric columns. Default is 0.}

\item{percent_decimals}{Decimal places of the percent columns. Default is 0.}
}
\value{
a gt-table
}
\description{
Knits table for one ore more variables
}
\examples{
WoJ |> knit_frequencies(reach, employment)
mtcars |> knit_frequencies(mpg, num_decimal = 1, percent_decimal = 1, cums = FALSE)

}
\seealso{
Other categorical: 
\code{\link{crosstab}()},
\code{\link{tab_frequencies}()}
}
\concept{categorical}
