% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_regression.R
\name{knit_regress_table}
\alias{knit_regress_table}
\title{Compute linear regression}
\usage{
knit_regress_table(
  x,
  digits = 2,
  R2adj = FALSE,
  CIs = TRUE,
  cap = NULL,
  B = B,
  LL = LL,
  UL = UL,
  beta_LL_compare = NULL,
  beta_UL_compare = NULL,
  beta_LL = NULL,
  beta_UL = NULL
)
}
\arguments{
\item{x}{a \link{tdcmm} model}

\item{digits}{the decimal digits}

\item{R2adj}{Show R^2adj in the table footnote. Default is FALSE.}

\item{CIs}{show the Confidence intervals or not. Default to TRUE.}

\item{cap}{Set the caption. Default is NULL, because in Quarto its better to set the tbl-cap in the chunk options}
}
\value{
a gt-table
}
\description{
Computes linear regression for all independent variables on the specified
dependent variable. Linear modeling of multiple independent variables uses
stepwise regression modeling. If specified, preconditions for
(multi-)collinearity and for homoscedasticity are checked.
}
\examples{
WoJ |> regress(autonomy_selection, ethics_1) |>
  knit_regress_table()
WoJ \%>\% regress(autonomy_selection, work_experience, trust_government) |>
  knit_regress_table(digits = 3, CIs = FALSE, cap = "Regression on Autonomy Selection")

}
