% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcmm_class.R
\name{model}
\alias{model}
\title{Access model(s) used to estimate output}
\usage{
model(x, ...)
}
\arguments{
\item{x}{\code{tdcmm} output}

\item{...}{other arguments}
}
\value{
A model object or a list of model objects
}
\description{
Returns model objects used to estimate \code{tdcmm} output.
}
