% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcmm_class.R
\name{new_tdcmm}
\alias{new_tdcmm}
\alias{is_tdcmm}
\title{\code{tdcmm} output constructor}
\usage{
new_tdcmm(x, func, data, model = NULL, checks = NULL, params = list())

is_tdcmm(x)
}
\arguments{
\item{x}{A \link[tibble:tibble]{tibble}.}

\item{func}{Function name called that produced this model.}

\item{data}{The \link[tibble:tibble]{tibble} that served as input to the function.}

\item{model}{A list of model object(s) used in preparation of the output.
Defaults to \code{NULL}. A single model should be wrapped in a list of length
\code{1}.}

\item{checks}{A list of assumption check object(s) used in preparation of the
output. Defaults to \code{NULL}.}

\item{params}{A named list of parameters originally passed to the call.
Defaults to an empty list.}
}
\description{
Creates a new \code{tdcmm} class output object.

The \code{tdcmm} class is a subclass of a \code{\link[tibble:tbl_df-class]{tbl_df}},
also know as a "tibble", used for augmenting output tibbles of \code{tidycomm}
functions with additional information. For output based on statistical tests,
the model object(s) estimated and any performed assumption checks.

\code{tdcmm} objects in \code{tidycomm} are further subclassed with individual classes
handling visualization and printing per "output" type.
}
\section{Functions}{
\itemize{
\item \code{is_tdcmm()}: Test for class \code{tdcmm}

}}
\keyword{internal}
