% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentage_labeller}
\alias{percentage_labeller}
\title{Helper function for labelling purposes}
\usage{
percentage_labeller(x)
}
\arguments{
\item{x}{share between 0 and 1}
}
\value{
a string with formatted \% (rounded and suffixed)
}
\description{
Helper function for labelling purposes
}
\concept{tdcmm visualize}
\keyword{internal}
