% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{recode_cat_scale}
\alias{recode_cat_scale}
\title{Recode one or more categorical scales.}
\usage{
recode_cat_scale(data, ..., assign = NULL, overwrite = FALSE, name = NULL)
}
\arguments{
\item{data}{A tibble or a tdcmm model.}

\item{...}{Variables to recode.}

\item{assign}{A named vector where names are the old values and values are
the new values to be assigned. The special name \code{.else} can be used to
provide a default value for unmatched cases.}

\item{overwrite}{Logical. If \code{TRUE}, it overwrites the original variable(s).
You cannot specify both 'name' and 'overwrite' parameters simultaneously.}

\item{name}{The name of the new variable(s). By default, this is the same
name as the provided variable(s) but suffixed with \verb{_rec}.}
}
\value{
A \link{tdcmm} model or a tibble.
}
\description{
Recode one or more categorical scales.
}
\examples{
WoJ \%>\%
recode_cat_scale(country,
assign = c("Germany" = 1, "Switzerland" = 2 , .else = 3), overwrite = TRUE)
WoJ \%>\%
recode_cat_scale(country,
assign = c("Germany" = "german", "Switzerland" = "swiss" , .else = "other"),
overwrite = TRUE)
WoJ \%>\%
recode_cat_scale(ethics_1, ethics_2,
assign = c(`1` = 5, `2` = 4, `3` = 3, `4` = 2, `5` = 1), overwrite = TRUE)
WoJ \%>\%
recode_cat_scale(ethics_1, ethics_2,
assign = c(`1` = "very low", `2` = "low", `3` = "medium", `4` = "high", `5` = "very high"),
overwrite = TRUE)
WoJ \%>\%
dplyr::select(temp_contract) \%>\% recode_cat_scale(temp_contract,
assign = c(`Permanent` = "P", `Temporary` = "T", .else = "E"))
}
\seealso{
Other scaling: 
\code{\link{center_scale}()},
\code{\link{dummify_scale}()},
\code{\link{minmax_scale}()},
\code{\link{reverse_scale}()},
\code{\link{setna_scale}()},
\code{\link{z_scale}()}
}
\concept{scaling}
