% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{recode_scale}
\alias{recode_scale}
\title{Recode one or more numeric, continuous scales into categories.}
\usage{
recode_scale(data, ..., breaks, labels, name = NULL, overwrite = FALSE)
}
\arguments{
\item{data}{A tibble or a tdcmm model.}

\item{...}{Variables to recode as factor variables in categories. If none are provided, all numeric columns in
the data will be recoded.}

\item{breaks}{A vector of numeric values specifying the breaks.}

\item{labels}{A vector of string labels for each interval.}

\item{name}{The name of the new variable(s). By default, this is the same
name as the provided variable(s) but suffixed with \verb{_cat}.}

\item{overwrite}{Logical. If \code{TRUE}, it overwrites the original variable(s).
You cannot specify both 'name' and 'overwrite' parameters simultaneously.}
}
\value{
A \link{tdcmm} model or a tibble.
}
\description{
This recodingsallows for the recoding of one or more numeric variables
into categorical variables based on provided breaks and labels.
If no variables are specified, all numeric columns will be recoded.
}
\concept{scaling
#' @examples
WoJ \%>\%
recode_scale(autonomy_emphasis, autonomy_selection, breaks = c(0, 1, 4, Inf),
labels = c("Low", "Medium", "High"), overwrite = TRUE)
WoJ \%>\%
dplyr::select(autonomy_emphasis) \%>\%
recode_scale(autonomy_emphasis, breaks = c(0, 1, 4, Inf),
labels = c("Low", "Medium", "High"), name = "new_na_autonomy")}
