% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcmm_class.R
\name{tdcmm-class}
\alias{tdcmm-class}
\alias{tdcmm}
\title{\code{tdcmm} class}
\description{
The \code{tdcmm} class is a subclass of a \code{\link[tibble:tbl_df-class]{tbl_df}},
also know as a "tibble", used for augmenting output tibbles of \code{tidycomm}
functions with additional information. For output based on statistical tests,
the model object(s) estimated and any performed assumption checks.

\code{tdcmm} objects in \code{tidycomm} are further subclassed with individual classes
handling visualization and printing per "output" type.
}
