% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{to_correlation_matrix}
\alias{to_correlation_matrix}
\title{Create correlation matrix}
\usage{
to_correlation_matrix(data, verbose = FALSE)
}
\arguments{
\item{data}{a \link{tdcmm} model returned from \code{\link{correlate}}}

\item{verbose}{A logical, defaulted to \code{FALSE}. Only applicable when
correlating two variables. If set to \code{TRUE}, the function outputs information
regarding the sample size.}
}
\value{
a \link{tdcmm} model
}
\description{
Turns the tibble exported from \code{\link{correlate}} into a correlation
matrix.
}
\examples{
WoJ \%>\% correlate() \%>\% to_correlation_matrix()

}
\concept{correlation}
