% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr.R
\name{unit_coder_matrix}
\alias{unit_coder_matrix}
\title{Generate units-coders matrix}
\usage{
unit_coder_matrix(data, unit_var, coder_var, test_var)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{unit_var}{Variable with unit identifiers}

\item{coder_var}{Variable with coder identifiers}

\item{test_var}{Variable to compute estimates for}
}
\description{
Generates a units-coders matrix for a test variable
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}},
  \code{\link{icr_brennan_prediger}},
  \code{\link{icr_cohens_kappa}},
  \code{\link{icr_fleiss_kappa}},
  \code{\link{icr_holstis_CR}},
  \code{\link{icr_kripp_alpha}}, \code{\link{test_icr}}
}
\concept{intercoder reliability}
