% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kripp_alpha.R
\name{values_in_unit}
\alias{values_in_unit}
\title{Count values in unit}
\usage{
values_in_unit(u, vals)
}
\arguments{
\item{u}{Unit of a unit-coders matrix}

\item{vals}{Values to count}
}
\description{
Counts occurences of values in a single unit of a units-coders matrix.
}
