% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcmm_class.R
\name{visualize}
\alias{visualize}
\title{Access visualization used to estimate output}
\usage{
visualize(x, ...)
}
\arguments{
\item{x}{\code{tdcmm} output}

\item{...}{other arguments}
}
\value{
A \link{ggplot2} object
}
\description{
Returns \link{ggplot2} visualization appropriate to respective \code{tdcmm} model
(see list below). Returns \code{NULL} (and a warning) if not visualization has
been implemented for the particular model.
}
