% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WoJ}
\alias{WoJ}
\title{Worlds of Journalism sample data}
\format{
A data frame with 1200 rows and 15 variables:
\describe{
\item{country}{Country of residence}
\item{reach}{Reach of medium}
\item{employment}{Current employment situation}
\item{temp_contract}{Type of contract (if current employment situation is
either full-time or part-time}
\item{autonomy_selection}{Autonomy in news story selection, scale from 1
(\emph{no freedom at all}) to 5 (\emph{complete freedom})}
\item{autonomy_emphasis}{Autonomy in news story emphasis, scale from 1
(\emph{no freedom at all}) to 5 (\emph{complete freedom})}
\item{ethics_1}{Agreement with statement "Journalists should always adhere
to codes of professional ethics, regardless of situation and context",
scale from 1 (\emph{strongly disagree}) to 5 (\emph{strongly agree})
(\emph{reverse-coded!})}
\item{ethics_2}{Agreement with statement "What is ethical in journalism
depends on the specific situation.",
scale from 1 (\emph{strongly disagree}) to 5 (\emph{strongly agree})}
\item{ethics_3}{Agreement with statement "What is ethical in journalism
is a matter of personal judgment.",
scale from 1 (\emph{strongly disagree}) to 5 (\emph{strongly agree})}
\item{ethics_4}{Agreement with statement "It is acceptable to set aside
moral standards if extraordinary circumstances require it.",
scale from 1 (\emph{strongly disagree}) to 5 (\emph{strongly agree})}
\item{work_experience}{Work experience as a journalist in years}
\item{trust_parliament}{Trust placed in parliament,
scale from 1 (\emph{no trust at all}) to 5 (\emph{complete trust})}
\item{trust_government}{Trust placed in government,
scale from 1 (\emph{no trust at all}) to 5 (\emph{complete trust})}
\item{trust_parties}{Trust placed in parties,
scale from 1 (\emph{no trust at all}) to 5 (\emph{complete trust})}
\item{trust_politicians}{Trust placed in politicians in general,
scale from 1 (\emph{no trust at all}) to 5 (\emph{complete trust})}
}
}
\source{
\samp{https://worldsofjournalism.org/data/data-and-key-tables-2012-2016}
}
\usage{
WoJ
}
\description{
A subset of data from the \href{https://worldsofjournalism.org/}{Worlds of Journalism}
2012-16 study containing survey data of 1,200 journalists from five European
countries.
}
\keyword{datasets}
