% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{z_scale}
\alias{z_scale}
\title{Z-standardize a numeric continuous scale}
\usage{
z_scale(
  data,
  scale_var,
  name = paste0(as_label(expr({
     {
         scale_var
     }
 })), "_z")
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} or a \link{tdcmm} model}

\item{scale_var}{a numeric variable for which the scale gets z-standardized}

\item{name}{the name of the new variable. By default, this is the same name
as \code{scale_var} but suffixed with \verb{_z}.}
}
\value{
a \link{tdcmm} model
}
\description{
Subtracts the mean from each individual data point and then divides the
result by the standard deviation. Eventually, a z-standardized scale centers
at a mean of 0.0 and obtains a standard deviation of 1.0. It becomes thus
comparable to other z-standardized distributions.
}
\examples{
WoJ \%>\% z_scale(autonomy_emphasis)
WoJ \%>\% z_scale(autonomy_emphasis, name = "my_zstdized_variable")

WoJ \%>\%
  z_scale(autonomy_emphasis) \%>\%
  tab_frequencies(autonomy_emphasis, autonomy_emphasis_z)
}
\seealso{
Other scaling: 
\code{\link{center_scale}()},
\code{\link{change_scale}()},
\code{\link{reverse_scale}()}
}
\concept{scaling}
