% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{compute_reliability}
\alias{compute_reliability}
\title{Compute reliability}
\usage{
compute_reliability(data, ..., type = "alpha", interval.type = NULL,
  bootstrap.samples = NULL, conf.level = NULL)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{...}{Variables to compute reliability estimate for}

\item{type}{Type of reliability estimate. See \code{\link[MBESS]{ci.reliability}}}

\item{interval.type}{Type of reliability estimate confidence interval.
See \code{\link[MBESS]{ci.reliability}}}

\item{bootstrap.samples}{Number of bootstrap samples for CI calculation.
See \code{\link[MBESS]{ci.reliability}}}

\item{conf.level}{Confidence level for estimate CI.
See \code{\link[MBESS]{ci.reliability}}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Computes reliability estimates of index/scale variables.
Provides a wrapper for \code{\link[MBESS]{ci.reliability}}.
}
\seealso{
Other reliability: \code{\link{get_reliability}}
}
\concept{reliability}
