% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlate}
\alias{correlate}
\title{Compute correlation coefficients}
\usage{
correlate(data, ..., method = "pearson", partial = FALSE, with = NULL)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} or a \link{tdcmm} model}

\item{...}{Variables to compute correlations for (column names). Leave empty
to compute for all numeric variables in data.}

\item{method}{a character string indicating which correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or "spearman"}

\item{partial}{Either a logical value or a character string. When set
to \code{TRUE}, the function computes the pairwise partial correlation for three
variables. If set to a character, it represents the variable to control for.
Defaults to \code{FALSE}.}

\item{with}{A character string specifying a focus variable to correlate all
other variables with. If used, \code{partial} must be set to \code{FALSE} (default).}
}
\value{
a \link{tdcmm} model
}
\description{
Computes correlation coefficients for all combinations of the specified
variables. If no variables are specified, all numeric (integer or double)
variables are used.
}
\examples{
WoJ \%>\% correlate(ethics_1, ethics_2, ethics_3)
WoJ \%>\% correlate()
WoJ \%>\% correlate(autonomy_selection, autonomy_emphasis, work_experience, partial = TRUE)

}
\concept{correlations}
