% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe variables}
\usage{
describe(data, ..., na.rm = TRUE)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{...}{Variables to describe (column names). Leave empty to describe all
numeric variables in data.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped
before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Describe variables by several measures of
central tendency and variability. If no variables are specified,
all numeric (integer or double) variables are described.
}
\examples{
iris \%>\% describe()
mtcars \%>\% describe(mpg, am, cyl)

}
\concept{descriptives}
