% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{dummify_scale}
\alias{dummify_scale}
\title{Convert categorical scales to dummy variables}
\usage{
dummify_scale(data, ..., overwrite = FALSE)
}
\arguments{
\item{data}{A \link[tibble:tibble-package]{tibble} or a \link{tdcmm} model.}

\item{...}{Categorical variables to be transformed into dummy variables.}

\item{overwrite}{Logical. If \code{TRUE}, it overwrites the original variable(s) with the dummy variables. If \code{FALSE} (default), new variables are created.}
}
\value{
A \link{tdcmm} model with the dummy variables appended.
}
\description{
This function transforms specified categorical variables into dummy variables.
Each level of the categorical variable is represented by a new dummy variable.
These new dummy variables are appended to the original data frame.
}
\examples{
WoJ \%>\% dplyr::select(temp_contract) \%>\% dummify_scale(temp_contract)
WoJ \%>\% dummify_scale(overwrite = TRUE)
}
\seealso{
Other scaling: 
\code{\link{center_scale}()},
\code{\link{minmax_scale}()},
\code{\link{recode_cat_scale}()},
\code{\link{reverse_scale}()},
\code{\link{setna_scale}()},
\code{\link{z_scale}()}
}
\concept{scaling}
