% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{get_reliability}
\alias{get_reliability}
\title{Get reliability estimates of index variables}
\usage{
get_reliability(data, ..., type = "alpha", interval.type = NULL,
  bootstrap.samples = NULL, conf.level = NULL, progress = FALSE)
}
\description{
Get reliability estimates of index variables created with \code{\link{add_index}}.
}
