% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab_vars}
\alias{grab_vars}
\title{Keep existing variables or get all numeric variables}
\usage{
grab_vars(data, vars, alternative = "numeric")
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{vars}{Variables passed to function with \code{...}, wrapped in \code{rlang::enquos}}

\item{alternative}{Which variables to grab alternatively if no variables were
specified in the function call. Defaults to "numeric".}
}
\value{
Variables as symbols
}
\description{
Keeps existing variables if they were specified in the function call or
gets an alternative selection of variables if none were specified.
}
