% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_agreement.R
\name{icr_agreement}
\alias{icr_agreement}
\title{Compute simple percent agreement}
\usage{
icr_agreement(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes simple percent agreement for a units-coders matrix
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_brennan_prediger}},
  \code{\link{icr_cohens_kappa}},
  \code{\link{icr_fleiss_kappa}},
  \code{\link{icr_holstis_CR}},
  \code{\link{icr_kripp_alpha}}, \code{\link{test_icr}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
