% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_kappa.R
\name{icr_cohens_kappa}
\alias{icr_cohens_kappa}
\title{Compute Cohen's Kappa}
\usage{
icr_cohens_kappa(ucm)
}
\arguments{
\item{ucm}{Units-coders matrix}
}
\description{
Computes Cohen's Kappa.
}
\references{
Cohen, J. (1960). A coefficient of agreement for nominal scales.
Educational and Psychological Measurement, 20(1), 37-46.
https://doi.org/10.1177/001316446002000104
}
\seealso{
Other intercoder reliability: \code{\link{compute_icr}},
  \code{\link{icr_agreement}},
  \code{\link{icr_brennan_prediger}},
  \code{\link{icr_fleiss_kappa}},
  \code{\link{icr_holstis_CR}},
  \code{\link{icr_kripp_alpha}}, \code{\link{test_icr}},
  \code{\link{unit_coder_matrix}}
}
\concept{intercoder reliability}
