% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_frequencies.R
\name{knit_frequencies}
\alias{knit_frequencies}
\title{knit frequencies table variables}
\usage{
knit_frequencies(
  data,
  ...,
  width_pct = 80,
  weight = NULL,
  num_decimal = 0,
  percent_decimal = 0,
  name_n = "N",
  name_total_percent = "Prozent",
  name_valid_percent = "Valide \%",
  name_cum_n = "Kum n",
  name_cum_percent = "Kum \%",
  cums = TRUE
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble}}

\item{...}{List of variables.}

\item{weight}{Weighting variable. Default = NULL, means all weights are 1.}

\item{num_decimal}{Decimal places of the numeric columns. Default is 0.}

\item{percent_decimal}{Decimal places of the percent columns. Default is 0.}

\item{name_n}{Rename Column n. Default is "N".}

\item{name_total_percent}{Rename total percent. Default is "Prozent".}

\item{name_valid_percent}{Rename valid percent. Default is "N".}

\item{name_cum_n}{Rename cum n. Default is "Kum n".}

\item{name_cum_percent}{Rename cum n. Default is "Kum \%".}

\item{cums}{Produce columns for cumulative counts and percents.}
}
\value{
a gt-table
}
\description{
Knits table for one ore more variables
}
\examples{
WoJ |> knit_frequencies(reach, employment)
mtcars |> knit_frequencies(mpg, num_decimal = 1, percent_decimal = 1, cums = FALSE)

}
\seealso{
Other categorical: 
\code{\link{crosstab}()},
\code{\link{tab_frequencies}()}
}
\concept{categorical}
