% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{reverse_scale}
\alias{reverse_scale}
\title{Reverse a numeric continuous scale}
\usage{
reverse_scale(
  data,
  scale_var,
  name = paste0(as_label(expr({
     {
         scale_var
     }
 })), "_rev")
)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} or a \link{tdcmm} model}

\item{scale_var}{a numeric variable for which the scale gets reversed}

\item{name}{the name of the new variable. By default, this is the same name
as the \code{scale_var} but suffixed with \verb{_rev}.}
}
\value{
a \link{tdcmm} model
}
\description{
Reverses a continuous scale into a new variable. A 5-1 scale thus turns into
a 1-5 scale. Missing values are retained. This function really just subtracts
each value from the maximum value which was incremented by one.
}
\examples{
WoJ \%>\% reverse_scale(autonomy_emphasis)
WoJ \%>\% reverse_scale(autonomy_emphasis, name = "my_reversed_variable")

WoJ \%>\%
  reverse_scale(autonomy_emphasis) \%>\%
  tab_frequencies(autonomy_emphasis, autonomy_emphasis_rev)
}
\seealso{
Other scaling: 
\code{\link{center_scale}()},
\code{\link{change_scale}()},
\code{\link{z_scale}()}
}
\concept{scaling}
