% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcmm_class.R
\name{tdcmm-class}
\alias{tdcmm-class}
\alias{tdcmm}
\title{\code{tdcmm} class}
\description{
The tdcmm class is a specialized type of data frame that is a
subclass of \link[tibble:tbl_df-class]{tbl_df}, also known as a "tibble". It is
designed to be used with \code{tidycomm} functions to provide additional
information and context to the output tibbles generated by these functions.
This subclass specifically augments output based on statistical tests, by
including the relevant model object(s) estimated during the analysis, as
well as any performed assumption checks.

\code{tdcmm} objects in \code{tidycomm} are further subclassed with individual classes
handling visualization and printing per "output" type.
}
