% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{to_correlation_matrix}
\alias{to_correlation_matrix}
\title{Create correlation matrix}
\usage{
to_correlation_matrix(data)
}
\arguments{
\item{data}{a \link[tibble:tibble-package]{tibble} returned from \code{\link{correlate}}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Turns the tibble exported from \code{\link{correlate}} into a correlation
matrix.
}
\examples{
WoJ \%>\% correlate() \%>\% to_correlation_matrix()

}
\concept{correlation}
